test: split compact context-budget suite below the max-lines cap
commands-compact.test.ts crossed the 1000-counted-line test cap when
#123640 added meter-refresh coverage, failing check-lint-core-4 on every
PR's merge-ref CI. Move the coherent context-budget block into a themed
sibling and share the fixtures via a test-support module; test count
parity 38 = 38.
