+++ src/auto-reply/reply/commands-compact.test-support.ts
@@ -1,0 +1,41 @@
+// Shared fixtures for the compact-command test suites; the main suite sits at
+// the max-lines cap, so themed siblings reuse these instead of duplicating.
+import { vi } from "vitest";
+import type { OpenClawConfig } from "../../config/config.js";
+import { compactEmbeddedAgentSession } from "./commands-compact.runtime.js";
+import type { HandleCommandsParams } from "./commands-types.js";
+
+export function buildCompactParams(
+  commandBodyNormalized: string,
+  cfg: OpenClawConfig,
+): HandleCommandsParams {
+  return {
+    cfg,
+    ctx: {
+      Provider: "whatsapp",
+      Surface: "whatsapp",
+      CommandSource: "text",
+      CommandBody: commandBodyNormalized,
+      commandText: commandBodyNormalized,
+    },
+    command: {
+      commandBodyNormalized,
+      isAuthorizedSender: true,
+      senderIsOwner: false,
+      senderId: "owner",
+      channel: "whatsapp",
+      ownerList: [],
+    },
+    sessionKey: "agent:main:main",
+    sessionStore: {},
+    resolveDefaultThinkingLevel: async () => "medium",
+  } as unknown as HandleCommandsParams;
+}
+
+export function requireCompactEmbeddedAgentSessionCall(index = 0) {
+  const call = vi.mocked(compactEmbeddedAgentSession).mock.calls[index]?.[0];
+  if (!call) {
+    throw new Error(`compactEmbeddedAgentSession call ${index} missing`);
+  }
+  return call;
+}
